renamed folder hooks on context
